backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
DateTime==6.0
Deprecated==1.3.1
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
fonttools==4.61.1
frozendict==2.4.7
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
multipledispatch==1.0.0
multitasking==0.0.12
numpy==2.3.5
oauthlib==3.3.1
openai==2.13.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.1
protobuf==6.33.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pykrx==1.0.51
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
selenium==4.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
sqlparse==0.5.4
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
wsproto==1.3.2
xlrd==2.0.2
yfinance==1.0
zope.interface==8.1.1
